samples/requirements.txt
jsonschema>=2.5.0
python-dateutil==2.6.0
pycurl>=7.19.5.1
